requirements.txt
paho-mqtt==2.0.0
pip==24.0
setuptools==69.1.1
